+++ CHANGELOG.rst
@@ -2,6 +2,9 @@ Changelog
 =========
 
 
+- Include execption class name in error message returned by ``As``.
+
+
 v0.2.0 (2018-07-25)
 -------------------